hide add-intermediate-waypoints button upon hitting the limit
